usb: dwc3: core: assert cru reset before power on phy
This patch prevent phy poweron failed.
log:
[    8.794869] phy phy-ff800000.phy.12: phy poweron failed --> -110
[    8.796743] dwc3 fe900000.usb: error -ETIMEDOUT: failed to initialize core
[    8.798810] dwc3: probe of fe900000.usb failed with error -110

The reason for "phy power on fail" could be that the initial state
of the phy's pipe is incorrect, and the pipe state is influenced by
the USB controller. For example, if the USB host is started during
the u-boot phase, it may affect the pipe state. Reset the USB controller
before powering on the phy to ensure the controller does not affect
the pipe state.

Signed-off-by: William Wu <william.wu@rock-chips.com>
Change-Id: I9f47b1c8e1b521cb82af6acad6de9b4e1a078f41